[Core] Split Config and Script IDL headers
* Moved Config and Script class declarations into dedicated generated module headers
* [IDL] Added forward class declarations and reference tracking for split headers
* [Doc] Regenerated module and class XML for the split definitions
