+++ docs/xml/modules/classes/bitmap.xml
@@ -943,6 +943,20 @@ unpackAlpha(Colour)
       <const name="SRGB">The default colour-space is sRGB.</const>
     </constants>
 
+    <constants lookup="MEM" comment="Memory types used by AllocMemory().  The lower 16 bits are stored with allocated blocks, the upper 16 bits are function-relative only.">
+      <const name="AUDIO">Identifies memory located on an audio device, such as a sound card.</const>
+      <const name="CODE">Used for memory that can execute code.</const>
+      <const name="DATA">The default type, <code>DATA</code>, is used to indicate a standard memory allocation from system RAM.</const>
+      <const name="NO_CLEAR">Do not clear the memory on allocation (saves time).</const>
+      <const name="READ">The memory is explicitly marked as readable.</const>
+      <const name="READ_WRITE">Synonym for <code>READ | WRITE</code></const>
+      <const name="STRING">Identifies the memory content as a null terminated string, which aids diagnostics.</const>
+      <const name="TEXTURE">Identifies non-displayable video memory, e.g. texture graphics.</const>
+      <const name="UNTRACKED">Prevents the memory block from being tracked back to the current context.</const>
+      <const name="VIDEO">Identifies video memory, such as the frame buffer.</const>
+      <const name="WRITE">The memory is explicitly marked as writeable.</const>
+    </constants>
+
   </types>
   <structs>
     <struct name="ColourFormat" typeName="COLOURFORMAT">
