+++ docs/xml/modules/classes/http.xml
@@ -498,6 +498,21 @@ http.acActivate()
 
   </fields>
   <types>
+    <constants lookup="DATA" comment="Data codes">
+      <const name="AUDIO">Audio file data, recognised by the Sound class</const>
+      <const name="CONTENT">Document content (between XML tags) - sent by document objects only</const>
+      <const name="DEVICE_INPUT">Device activity</const>
+      <const name="FILE">File location (the data will reflect the complete file path)</const>
+      <const name="IMAGE">Image file data, recognised by the Image class</const>
+      <const name="INPUT_READY">Device input that has been transformed into user input</const>
+      <const name="RAW">Raw unprocessed data</const>
+      <const name="RECEIPT">Receipt for item data, in response to an earlier request</const>
+      <const name="RECORD">Database record</const>
+      <const name="REQUEST">Make a request for item data</const>
+      <const name="TEXT">Standard ASCII text</const>
+      <const name="XML">Markup based text data.  NOTE - For clipboard data, the top-level encapsulating tag must declare the type of XML, e.g. 'html', 'ripple'.  For plain XML, use 'xml'</const>
+    </constants>
+
     <constants lookup="HGS" comment="Options for defining an HTTP object's state.">
       <const name="AUTHENTICATED">Following successful authentication, this state value will be set.</const>
       <const name="AUTHENTICATING">If authentication is requested from the server, the HTTP object will enter this state.</const>
