+++ docs/xml/modules/network.xml
@@ -169,64 +169,42 @@ if (!StrToAddress(&quot;127.0.0.1&quot;, &amp;addr)) {
       <const name="V6"/>
     </constants>
 
-    <constants lookup="NLF" comment="Options for NetLookup">
-      <const name="NO_CACHE">Contact the name service and do not use the local DNS cache.</const>
+    <constants lookup="JET" comment="JET constants are documented in GetInputEvent()">
+      <const name="ABS_XY">The X, Y values are defined as absolute coordinates, relative to the top-left of the display.</const>
+      <const name="BUTTON_1">Left mouse button; XBox A button, PS square button.  Value is pressure sensitive, ranging between 0 - 1.0 (0 is released, 1.0 is fully depressed).</const>
+      <const name="BUTTON_10">Non-specific button assignment.</const>
+      <const name="BUTTON_2">Right mouse button; XBox X button, PS cross button.</const>
+      <const name="BUTTON_3">Middle mouse button; XBox Y button, PS triangle.</const>
+      <const name="BUTTON_4">Alt. mouse button 1; XBox B button, PS circle.</const>
+      <const name="BUTTON_5">Alt. mouse button 2.</const>
+      <const name="BUTTON_6">Non-specific button assignment.</const>
+      <const name="BUTTON_7">Non-specific button assignment.</const>
+      <const name="BUTTON_8">Non-specific button assignment.</const>
+      <const name="BUTTON_9">Non-specific button assignment.</const>
+      <const name="CROSSED_IN">This message is sent by the input system when the mouse pointer enters an area for the first time.  The message value refers to the object ID of the container being monitored for movement.</const>
+      <const name="CROSSED_OUT">This message is sent by the input system when the mouse pointer leaves an area.  The message value refers to the object ID of the container being monitored for movement.</const>
+      <const name="DEVICE_TILT_XY">Controller tilted on the X/Y axis.  Value indicates angle, -ve = left, +ve = right</const>
+      <const name="DEVICE_TILT_Z">Controller is rising or falling.  Value expressed as 'speed',</const>
+      <const name="DISPLAY_EDGE">Recently supplied input occurred at the edge of the display.</const>
+      <const name="PEN_TILT_XY">Pen tilt angle.  0 is pointing down directly with nib at bottom, 0.5 is 90 degrees, 1.0 is inversed (eraser at bottom)</const>
+      <const name="PRESSURE">Amount of pressure applied, ranges from 0 (none) to 1.0 (normal) and possibly higher if user presses hard enough</const>
+      <const name="WHEEL">Mouse wheel rotation - the value generally reflects the number of 'clicks' rotated on the wheel.</const>
+      <const name="WHEEL_TILT">Some mouse wheels can be tilted to the left or right.  Ranges from -1.0 to +1.0</const>
     </constants>
 
-    <constants lookup="NSF" comment="NetSocket options">
-      <const name="BROADCAST">Enable broadcast (UDP only).</const>
-      <const name="DISABLE_SERVER_VERIFY">Disable SSL server certificate verification for client sockets (testing only).</const>
-      <const name="KEEP_ALIVE">Enable TCP keep-alive probes using operating system defaults.</const>
-      <const name="LOG_ALL">Print extra log messages.</const>
-      <const name="MULTI_CONNECT">Allow multiple connections from the same IP (NetServer only).</const>
-      <const name="SSL">Use Secure Sockets Layer for all communication.</const>
-      <const name="SYNCHRONOUS">Use synchronous (blocking) network calls.</const>
-      <const name="UDP">Use UDP (connectionless datagram protocol) instead of TCP.</const>
-    </constants>
-
-    <constants lookup="NTC" comment="NetSocket states">
-      <const name="CONNECTED">There is an active connection at present.</const>
-      <const name="CONNECTING">A connection is being established.</const>
-      <const name="DISCONNECTED">There is no connection.</const>
-      <const name="HANDSHAKING">An SSL connection is being established.</const>
-      <const name="MULTISTATE">A NetServer reports MULTISTATE because accepted ClientSocket objects track their own states.</const>
-      <const name="RESOLVING">The host name is being resolved.</const>
-    </constants>
-
-    <constants lookup="SCV" comment="These error codes for certificate validation match the OpenSSL error codes (X509 definitions)">
-      <const name="AKID_ISSUER_SERIAL_MISMATCH"/>
-      <const name="AKID_SKID_MISMATCH"/>
-      <const name="APPLICATION_VERIFICATION"/>
-      <const name="CERT_CHAIN_TOO_LONG"/>
-      <const name="CERT_HAS_EXPIRED"/>
-      <const name="CERT_NOT_YET_VALID"/>
-      <const name="CERT_REJECTED"/>
-      <const name="CERT_REVOKED"/>
-      <const name="CERT_SIGNATURE_FAILURE"/>
-      <const name="CERT_UNTRUSTED"/>
-      <const name="CRL_HAS_EXPIRED"/>
-      <const name="CRL_NOT_YET_VALID"/>
-      <const name="CRL_SIGNATURE_FAILURE"/>
-      <const name="DEPTH_ZERO_SELF_SIGNED_CERT"/>
-      <const name="ERROR_IN_CERT_NOT_AFTER_FIELD"/>
-      <const name="ERROR_IN_CERT_NOT_BEFORE_FIELD"/>
-      <const name="ERROR_IN_CRL_LAST_UPDATE_FIELD"/>
-      <const name="ERROR_IN_CRL_NEXT_UPDATE_FIELD"/>
-      <const name="INVALID_CA"/>
-      <const name="INVALID_PURPOSE"/>
-      <const name="KEYUSAGE_NO_CERTSIGN"/>
-      <const name="OK"/>
-      <const name="OUT_OF_MEM"/>
-      <const name="PATH_LENGTH_EXCEEDED"/>
-      <const name="SELF_SIGNED_CERT_IN_CHAIN"/>
-      <const name="SUBJECT_ISSUER_MISMATCH"/>
-      <const name="UNABLE_TO_DECODE_ISSUER_PUBLIC_KEY"/>
-      <const name="UNABLE_TO_DECRYPT_CERT_SIGNATURE"/>
-      <const name="UNABLE_TO_DECRYPT_CRL_SIGNATURE"/>
-      <const name="UNABLE_TO_GET_CRL"/>
-      <const name="UNABLE_TO_GET_ISSUER_CERT"/>
-      <const name="UNABLE_TO_GET_ISSUER_CERT_LOCALLY"/>
-      <const name="UNABLE_TO_VERIFY_LEAF_SIGNATURE"/>
+    <constants lookup="JTYPE" comment="JTYPE flags are used to categorise input types.">
+      <const name="ANALOG">Analog movement (ranging from -1.0 to 1.0)</const>
+      <const name="ANCHORED">Cursor has been anchored with LockCursor()</const>
+      <const name="BUTTON">Input is a physical button or switch</const>
+      <const name="CROSSING">Crossing events manage the entering and leaving of an area.</const>
+      <const name="DBL_CLICK">Set by the input system if the <code>Type</code> is a button and the button has been clicked in quick succession so as to be classed as a double-click.</const>
+      <const name="DIGITAL">D-Pad or digital joystick source (restricted to +/- 1)</const>
+      <const name="DRAGGED">Set if sufficient movement occurred between the original click and its point of release (usually requires a 3 or more pixel difference).</const>
+      <const name="DRAG_ITEM">This special flag is set by the input system if the pointer is click-dragging an object at the time of the event.</const>
+      <const name="EXT_MOVEMENT">Extended or indirect movement information.  This covers all types of movement that are unconnected to coordinate positioning -  mouse wheel movement and pen tilt are two such examples.</const>
+      <const name="MOVEMENT">X/Y coordinate movement only. Movement such as the wheel mouse spinning is not covered by this type as it does not influence the coordinate system.</const>
+      <const name="REPEATED">Input is a repeated entry (i.e. user is holding down a button and a repetition timer is being triggered)</const>
+      <const name="SECONDARY">Indicates to the receiver of this message that it is not the primary/original recipient</const>
     </constants>
 
   </types>
