+++ docs/xml/modules/tiri.xml
@@ -47,24 +47,42 @@
   </function>
 
   <types>
-    <constants lookup="JOF" comment="JIT behaviour options">
-      <const name="ALL_TIPS">Include all tips and style suggestions.</const>
-      <const name="DIAGNOSE">Cancels aborting on error so that the full script is parsed.  Enables deeper log tracing.</const>
-      <const name="DISABLE_JIT">Disable the JIT compiler.</const>
-      <const name="DUMP_BYTECODE">Print a disassembly of bytecode instructions after parsing.</const>
-      <const name="PROFILE">Use timers to profile JIT parsing.  Required to configure run-time profiling.</const>
-      <const name="TIPS">Filter tips down to performance and code quality issues.</const>
-      <const name="TOP_TIPS">Filter tips down to type safety and correctness issues.</const>
-      <const name="TRACE">Synonym for <code>TRACE_TOKENS | TRACE_EXPECT | TRACE_BOUNDARY | TRACE_OPERATORS | TRACE_REGISTERS | TRACE_CFG | TRACE_ASSIGNMENTS | TRACE_VALUE_CATEGORY | TRACE_TYPES</code></const>
-      <const name="TRACE_ASSIGNMENTS">Trace PreparedAssignment lifecycle and hazard handling.</const>
-      <const name="TRACE_BOUNDARY">Enable log messages for AST boundary crossing.</const>
-      <const name="TRACE_CFG">Trace ControlFlowGraph edge creation and patching.</const>
-      <const name="TRACE_EXPECT">Enable trace expectation messages during parsing.</const>
-      <const name="TRACE_OPERATORS">Trace OperatorEmitter method calls and decisions.</const>
-      <const name="TRACE_REGISTERS">Trace RegisterAllocator reservations and releases.</const>
-      <const name="TRACE_TOKENS">Trace the emission of tokens.</const>
-      <const name="TRACE_TYPES">Trace type inference and type checking decisions.</const>
-      <const name="TRACE_VALUE_CATEGORY">Trace ValueUse/ValueSlot/LValue operations.</const>
+    <constants lookup="JET" comment="JET constants are documented in GetInputEvent()">
+      <const name="ABS_XY">The X, Y values are defined as absolute coordinates, relative to the top-left of the display.</const>
+      <const name="BUTTON_1">Left mouse button; XBox A button, PS square button.  Value is pressure sensitive, ranging between 0 - 1.0 (0 is released, 1.0 is fully depressed).</const>
+      <const name="BUTTON_10">Non-specific button assignment.</const>
+      <const name="BUTTON_2">Right mouse button; XBox X button, PS cross button.</const>
+      <const name="BUTTON_3">Middle mouse button; XBox Y button, PS triangle.</const>
+      <const name="BUTTON_4">Alt. mouse button 1; XBox B button, PS circle.</const>
+      <const name="BUTTON_5">Alt. mouse button 2.</const>
+      <const name="BUTTON_6">Non-specific button assignment.</const>
+      <const name="BUTTON_7">Non-specific button assignment.</const>
+      <const name="BUTTON_8">Non-specific button assignment.</const>
+      <const name="BUTTON_9">Non-specific button assignment.</const>
+      <const name="CROSSED_IN">This message is sent by the input system when the mouse pointer enters an area for the first time.  The message value refers to the object ID of the container being monitored for movement.</const>
+      <const name="CROSSED_OUT">This message is sent by the input system when the mouse pointer leaves an area.  The message value refers to the object ID of the container being monitored for movement.</const>
+      <const name="DEVICE_TILT_XY">Controller tilted on the X/Y axis.  Value indicates angle, -ve = left, +ve = right</const>
+      <const name="DEVICE_TILT_Z">Controller is rising or falling.  Value expressed as 'speed',</const>
+      <const name="DISPLAY_EDGE">Recently supplied input occurred at the edge of the display.</const>
+      <const name="PEN_TILT_XY">Pen tilt angle.  0 is pointing down directly with nib at bottom, 0.5 is 90 degrees, 1.0 is inversed (eraser at bottom)</const>
+      <const name="PRESSURE">Amount of pressure applied, ranges from 0 (none) to 1.0 (normal) and possibly higher if user presses hard enough</const>
+      <const name="WHEEL">Mouse wheel rotation - the value generally reflects the number of 'clicks' rotated on the wheel.</const>
+      <const name="WHEEL_TILT">Some mouse wheels can be tilted to the left or right.  Ranges from -1.0 to +1.0</const>
+    </constants>
+
+    <constants lookup="JTYPE" comment="JTYPE flags are used to categorise input types.">
+      <const name="ANALOG">Analog movement (ranging from -1.0 to 1.0)</const>
+      <const name="ANCHORED">Cursor has been anchored with LockCursor()</const>
+      <const name="BUTTON">Input is a physical button or switch</const>
+      <const name="CROSSING">Crossing events manage the entering and leaving of an area.</const>
+      <const name="DBL_CLICK">Set by the input system if the <code>Type</code> is a button and the button has been clicked in quick succession so as to be classed as a double-click.</const>
+      <const name="DIGITAL">D-Pad or digital joystick source (restricted to +/- 1)</const>
+      <const name="DRAGGED">Set if sufficient movement occurred between the original click and its point of release (usually requires a 3 or more pixel difference).</const>
+      <const name="DRAG_ITEM">This special flag is set by the input system if the pointer is click-dragging an object at the time of the event.</const>
+      <const name="EXT_MOVEMENT">Extended or indirect movement information.  This covers all types of movement that are unconnected to coordinate positioning -  mouse wheel movement and pen tilt are two such examples.</const>
+      <const name="MOVEMENT">X/Y coordinate movement only. Movement such as the wheel mouse spinning is not covered by this type as it does not influence the coordinate system.</const>
+      <const name="REPEATED">Input is a repeated entry (i.e. user is holding down a button and a repetition timer is being triggered)</const>
+      <const name="SECONDARY">Indicates to the receiver of this message that it is not the primary/original recipient</const>
     </constants>
 
   </types>
