+++ include/kotuku/modules/core.h
@@ -72,8 +72,6 @@ constexpr int RESOURCE_ID_OFFSET = -1;
 class objMetaClass;
 class objStorageDevice;
 class objFile;
-class objConfig;
-class objScript;
 class objTask;
 class objThread;
 class objModule;
@@ -506,6 +504,18 @@ enum class ALIGN : uint32_t {
 };
 
 DEFINE_ENUM_FLAG_OPERATORS(ALIGN)
+
+// Flags for the Config class.
+
+enum class CNF : uint32_t {
+   NIL = 0,
+   STRIP_QUOTES = 0x00000001,
+   AUTO_SAVE = 0x00000002,
+   OPTIONAL_FILES = 0x00000004,
+   NEW = 0x00000008,
+};
+
+DEFINE_ENUM_FLAG_OPERATORS(CNF)
 
 // Script flags
 
@@ -1009,18 +1019,6 @@ enum class CLF : uint32_t {
 };
 
 DEFINE_ENUM_FLAG_OPERATORS(CLF)
-
-// Flags for the Config class.
-
-enum class CNF : uint32_t {
-   NIL = 0,
-   STRIP_QUOTES = 0x00000001,
-   AUTO_SAVE = 0x00000002,
-   OPTIONAL_FILES = 0x00000004,
-   NEW = 0x00000008,
-};
-
-DEFINE_ENUM_FLAG_OPERATORS(CNF)
 
 // Raw key codes
 
@@ -1849,27 +1847,6 @@ struct ClassRecord {
    std::string Description;  // File description
 };
 
-struct ScriptArg { // For use with sc::Exec
-   CSTRING Name;
-   uint32_t Type;
-   union {
-      APTR    Address;
-      int     Int;
-      int64_t Int64;
-      double  Double;
-   };
-
-   ScriptArg(CSTRING pName, OBJECTPTR pValue, uint32_t pType = FD_OBJECTPTR) : Name(pName), Type(pType), Address((APTR)pValue) { }
-   ScriptArg(CSTRING pName, std::string &pValue, uint32_t pType = FD_STRING) : Name(pName), Type(pType), Address((APTR)pValue.data()) { }
-   ScriptArg(CSTRING pName, const std::string &pValue, uint32_t pType = FD_STRING) : Name(pName), Type(pType), Address((APTR)pValue.data()) { }
-   ScriptArg(CSTRING pName, CSTRING pValue, uint32_t pType = FD_STRING) : Name(pName), Type(pType), Address((APTR)pValue) { }
-   ScriptArg(CSTRING pName, APTR pValue, uint32_t pType = FD_PTR) : Name(pName), Type(pType), Address(pValue) { }
-   ScriptArg(CSTRING pName, int pValue, uint32_t pType = FD_INT) : Name(pName), Type(pType), Int(pValue) { }
-   ScriptArg(CSTRING pName, uint32_t pValue, uint32_t pType = FD_INT) : Name(pName), Type(pType), Int(pValue) { }
-   ScriptArg(CSTRING pName, int64_t pValue, uint32_t pType = FD_INT64) : Name(pName), Type(pType), Int64(pValue) { }
-   ScriptArg(CSTRING pName, double pValue, uint32_t pType = FD_DOUBLE) : Name(pName), Type(pType), Double(pValue) { }
-};
-
 #ifdef KOTUKU_STATIC
 #define JUMPTABLE_CORE [[maybe_unused]] static struct CoreBase *CoreBase = nullptr;
 #else
@@ -2228,10 +2205,6 @@ inline ERR QueueAction(AC Action, OBJECTID ObjectID) {
    return QueueAction(Action, ObjectID, nullptr);
 }
 #endif
-
-typedef KEYVALUE ConfigKeys;
-typedef std::pair<std::string, ConfigKeys> ConfigGroup;
-typedef std::vector<ConfigGroup> ConfigGroups;
 
 namespace kt {
 
@@ -2972,464 +2945,6 @@ class objFile : public Object {
 
 };
 
-// Config class definition
-
-#define VER_CONFIG (1.000000)
-
-// Config methods
-
-namespace cfg {
-struct ReadValue { std::string_view Group; std::string_view Key; std::string_view Data; static const AC id = AC(-1); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct Set { std::string_view Group; std::string_view Key; std::string_view Data; static const AC id = AC(-2); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct WriteValue { std::string_view Group; std::string_view Key; std::string_view Data; static const AC id = AC(-3); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct DeleteKey { std::string_view Group; std::string_view Key; static const AC id = AC(-4); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct DeleteGroup { std::string_view Group; static const AC id = AC(-5); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct GetGroupFromIndex { int Index; std::string_view Group; static const AC id = AC(-6); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct SortByKey { std::string_view Key; int Descending; static const AC id = AC(-7); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct MergeFile { std::string_view Path; static const AC id = AC(-9); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct Merge { OBJECTPTR Source; static const AC id = AC(-10); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-
-} // namespace
-
-class objConfig : public Object {
-   public:
-   static constexpr CLASSID CLASS_ID = CLASSID::CONFIG;
-   static constexpr CSTRING CLASS_NAME = "Config";
-
-   using create = kt::Create<objConfig>;
-   objConfig(objMetaClass *pClass, OBJECTID pUID) noexcept : Object(pClass, pUID) {}
-
-   std::string Path;         // Set this field to the location of the source configuration file.
-   std::string KeyFilter;    // Set this field to enable key filtering.
-   std::string GroupFilter;  // Set this field to enable group filtering.
-   CNF Flags;                // Optional flags may be set here.
-   public:
-   ConfigGroups Groups; // Is a std::vector of std::pair and std::map values
-
-   inline ERR getGroups(ConfigGroups * &Groups) {
-      APTR ptr;
-      if (auto error = getData(ptr); error IS ERR::Okay) {
-         Groups = (ConfigGroups *)ptr;
-         return error;
-      }
-      else return error;
-   }
-
-   // For C++ only, these read variants avoid method calls for speed, but apply identical logic.
-
-   inline ERR read(std::string_view pGroup, std::string_view pKey, double &pValue) {
-      for (auto& [group, keys] : Groups) {
-         if ((!pGroup.empty()) and (group.compare(pGroup))) continue;
-         if (pKey.empty()) pValue = strtod(keys.cbegin()->second.c_str(), nullptr);
-         else if (auto it = keys.find(pKey); it != keys.end()) pValue = strtod(it->second.c_str(), nullptr);
-         else return ERR::Search;
-         return ERR::Okay;
-      }
-      return ERR::Search;
-   }
-
-   inline ERR read(std::string_view pGroup, std::string_view pKey, int &pValue) {
-      for (auto& [group, keys] : Groups) {
-         if ((!pGroup.empty()) and (group.compare(pGroup))) continue;
-         if (pKey.empty()) pValue = strtol(keys.cbegin()->second.c_str(), nullptr, 0);
-         else if (auto it = keys.find(pKey); it != keys.end()) pValue = strtol(it->second.c_str(), nullptr, 0);
-         else return ERR::Search;
-         return ERR::Okay;
-      }
-      return ERR::Search;
-   }
-
-   inline ERR read(std::string_view pGroup, std::string_view pKey, std::string &pValue) {
-      for (auto & [group, keys] : Groups) {
-         if ((!pGroup.empty()) and (group.compare(pGroup))) continue;
-         if (pKey.empty()) pValue = keys.cbegin()->second;
-         else if (auto it = keys.find(pKey); it != keys.end()) pValue = it->second;
-         else return ERR::Search;
-         return ERR::Okay;
-      }
-      return ERR::Search;
-   }
-
-   inline ERR write(std::string_view Group, std::string_view Key, std::string_view Value) {
-      for (auto& [group, keys] : Groups) {
-         if (!group.compare(Group)) {
-            if (auto it = keys.find(Key); it != keys.end()) {
-               it->second.assign(Value);
-            }
-            else keys.emplace(Key, Value);
-            return ERR::Okay;
-         }
-      }
-
-      auto &new_group = Groups.emplace_back();
-      new_group.first.assign(Group);
-      new_group.second.emplace(Key, Value);
-      return ERR::Okay;
-   }
-
-   // Action stubs
-
-   inline ERR clear() noexcept { return Action(AC::Clear, this, nullptr); }
-   inline ERR dataFeed(OBJECTPTR Object, DATA Datatype, const void *Buffer, int Size) noexcept {
-      struct acDataFeed args = { Object, Datatype, Buffer, Size };
-      return Action(AC::DataFeed, this, &args);
-   }
-   inline ERR flush() noexcept { return Action(AC::Flush, this, nullptr); }
-   inline ERR init() noexcept { return InitObject(this); }
-   inline ERR saveSettings() noexcept { return Action(AC::SaveSettings, this, nullptr); }
-   inline ERR saveToObject(OBJECTPTR Dest, CLASSID ClassID = CLASSID::NIL) noexcept {
-      struct acSaveToObject args = { Dest, { ClassID } };
-      return Action(AC::SaveToObject, this, &args);
-   }
-   inline ERR readValue(const std::string_view &Group, const std::string_view &Key, std::string_view * Data) noexcept {
-      struct cfg::ReadValue args = { Group, Key };
-      ERR error = Action(AC(-1), this, &args);
-      if (Data) *Data = args.Data;
-      return error;
-   }
-   inline ERR set(const std::string_view &Group, const std::string_view &Key, const std::string_view &Data) noexcept {
-      struct cfg::Set args = { Group, Key, Data };
-      return Action(AC(-2), this, &args);
-   }
-   inline ERR writeValue(const std::string_view &Group, const std::string_view &Key, const std::string_view &Data) noexcept {
-      struct cfg::WriteValue args = { Group, Key, Data };
-      return Action(AC(-3), this, &args);
-   }
-   inline ERR deleteKey(const std::string_view &Group, const std::string_view &Key) noexcept {
-      struct cfg::DeleteKey args = { Group, Key };
-      return Action(AC(-4), this, &args);
-   }
-   inline ERR deleteGroup(const std::string_view &Group) noexcept {
-      struct cfg::DeleteGroup args = { Group };
-      return Action(AC(-5), this, &args);
-   }
-   inline ERR getGroupFromIndex(int Index, std::string_view * Group) noexcept {
-      struct cfg::GetGroupFromIndex args = { Index };
-      ERR error = Action(AC(-6), this, &args);
-      if (Group) *Group = args.Group;
-      return error;
-   }
-   inline ERR sortByKey(const std::string_view &Key, int Descending) noexcept {
-      struct cfg::SortByKey args = { Key, Descending };
-      return Action(AC(-7), this, &args);
-   }
-   inline ERR mergeFile(const std::string_view &Path) noexcept {
-      struct cfg::MergeFile args = { Path };
-      return Action(AC(-9), this, &args);
-   }
-   inline ERR merge(OBJECTPTR Source) noexcept {
-      struct cfg::Merge args = { Source };
-      return Action(AC(-10), this, &args);
-   }
-
-   // Customised field getting
-
-   inline ERR getPath(std::string_view &Value) noexcept {
-      Value = this->Path;
-      return ERR::Okay;
-   }
-
-   inline ERR getKeyFilter(std::string_view &Value) noexcept {
-      auto field = &this->Class->Dictionary[5];
-      auto get_field = (ERR (*)(APTR, std::string_view &))field->GetValue;
-      return get_field(this, Value);
-   }
-
-   inline ERR getGroupFilter(std::string_view &Value) noexcept {
-      auto field = &this->Class->Dictionary[1];
-      auto get_field = (ERR (*)(APTR, std::string_view &))field->GetValue;
-      return get_field(this, Value);
-   }
-
-   inline ERR getFlags(CNF &Value) noexcept {
-      Value = this->Flags;
-      return ERR::Okay;
-   }
-
-   inline ERR getData(APTR &Value) noexcept {
-      auto field = &this->Class->Dictionary[8];
-      return field->GetValue(this, &Value);
-   }
-
-   inline ERR getTotalGroups(int &Value) noexcept {
-      auto field = &this->Class->Dictionary[6];
-      return field->GetValue(this, &Value);
-   }
-
-   inline ERR getTotalKeys(int &Value) noexcept {
-      auto field = &this->Class->Dictionary[11];
-      return field->GetValue(this, &Value);
-   }
-
-
-   // Customised field setting
-
-   inline ERR setPath(const std::string_view &Value) noexcept {
-      auto field = &this->Class->Dictionary[3];
-      return field->WriteValue(this, field, 0x00804300, &Value);
-   }
-
-   inline ERR setKeyFilter(const std::string_view &Value) noexcept {
-      auto field = &this->Class->Dictionary[5];
-      return field->WriteValue(this, field, 0x00904300, &Value);
-   }
-
-   inline ERR setGroupFilter(const std::string_view &Value) noexcept {
-      auto field = &this->Class->Dictionary[1];
-      return field->WriteValue(this, field, 0x00904300, &Value);
-   }
-
-   inline ERR setFlags(const CNF Value) noexcept {
-      this->Flags = Value;
-      return ERR::Okay;
-   }
-
-};
-
-// Script class definition
-
-#define VER_SCRIPT (1.000000)
-
-// Script methods
-
-namespace sc {
-struct Exec { std::string_view Procedure; const struct ScriptArg *Args; int TotalArgs; static const AC id = AC(-1); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct DerefProcedure { FUNCTION *Procedure; static const AC id = AC(-2); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct Callback { int64_t ProcedureID; const struct ScriptArg *Args; int TotalArgs; ERR Error; static const AC id = AC(-3); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct GetProcedureID { std::string_view Procedure; int64_t ProcedureID; static const AC id = AC(-4); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-struct DebugLog { std::string_view Options; std::string *Result; static const AC id = AC(-5); ERR call(OBJECTPTR Object) { return Action(id, Object, this); } };
-
-} // namespace
-
-class objScript : public Object {
-   public:
-   static constexpr CLASSID CLASS_ID = CLASSID::SCRIPT;
-   static constexpr CSTRING CLASS_NAME = "Script";
-
-   using create = kt::Create<objScript>;
-   objScript(objMetaClass *pClass, OBJECTID pUID) noexcept : Object(pClass, pUID) {}
-
-   std::string Procedure;              // Specifies a procedure name to be executed.
-   std::string CacheFile;              // Compilable script languages can be compiled to a cache file.
-   std::string Path;                   // The location of a script file to be loaded.
-   std::string ErrorMessage;           // A human readable error string may be declared here following a script execution failure.
-   std::string Statement;              // Scripts can be executed from any string passed into this field.
-   std::string WorkingPath;            // Defines the script's working path (folder).
-   kt::vector<std::string> Results;    // Stores multiple string results for languages that support this feature.
-   OBJECTID TargetID;                  // Reference to the default container that new script objects will be initialised to.
-   SCF      Flags;                     // Optional flags.
-   ERR      Error;                     // If a script fails during execution, an error code may be readable here.
-   int      CurrentLine;               // Indicates the current line being executed when in debug mode.
-   int      LineOffset;                // For debugging purposes, this value is added to any message referencing a line number.
-   public:
-   int64_t  ProcedureID;          // For callbacks
-   KEYVALUE Vars;                 // Global parameters
-   const ScriptArg *ProcArgs;     // Procedure args - applies during Exec
-   int      ActivationCount;      // Incremented every time the script is activated.
-   int      TotalArgs;            // Total number of ProcArgs
-   OBJECTID ScriptOwnerID;
-
-#ifdef KOTUKU_CXX_REUSES_BASE_TAIL_PADDING // Padding for the alignment of derived classes
-   uint8_t TailPadding[alignof(APTR) - sizeof(OBJECTID)];
-#endif
-
-   // Action stubs
-
-   inline ERR activate() noexcept { return Action(AC::Activate, this, nullptr); }
-   inline ERR dataFeed(OBJECTPTR Object, DATA Datatype, const void *Buffer, int Size) noexcept {
-      struct acDataFeed args = { Object, Datatype, Buffer, Size };
-      return Action(AC::DataFeed, this, &args);
-   }
-   inline ERR getKey(std::string_view Key, std::string &Value) noexcept {
-      struct acGetKey args = { Key, &Value };
-      auto error = Action(AC::GetKey, this, &args);
-      if (error != ERR::Okay) Value.clear();
-      return error;
-   }
-   inline ERR init() noexcept { return InitObject(this); }
-   inline ERR reset() noexcept { return Action(AC::Reset, this, nullptr); }
-   inline ERR acSetKey(std::string_view FieldName, std::string_view Value) noexcept {
-      struct acSetKey args = { FieldName, Value };
-      return Action(AC::SetKey, this, &args);
-   }
-   inline ERR exec(const std::string_view &Procedure, const struct ScriptArg * Args, int TotalArgs) noexcept {
-      struct sc::Exec args = { Procedure, Args, TotalArgs };
-      return Action(AC(-1), this, &args);
-   }
-   inline ERR derefProcedure(FUNCTION Procedure) noexcept {
-      struct sc::DerefProcedure args = { &Procedure };
-      return Action(AC(-2), this, &args);
-   }
-   inline ERR callback(int64_t ProcedureID, const struct ScriptArg * Args, int TotalArgs, ERR * Error) noexcept {
-      struct sc::Callback args = { ProcedureID, Args, TotalArgs, (ERR)0 };
-      ERR error = Action(AC(-3), this, &args);
-      if (Error) *Error = args.Error;
-      return error;
-   }
-   inline ERR getProcedureID(const std::string_view &Procedure, int64_t * ProcedureID) noexcept {
-      struct sc::GetProcedureID args = { Procedure, (int64_t)0 };
-      ERR error = Action(AC(-4), this, &args);
-      if (ProcedureID) *ProcedureID = args.ProcedureID;
-      return error;
-   }
-   inline ERR debugLog(const std::string_view &Options, std::string &Result) noexcept {
-      struct sc::DebugLog args = { Options, &Result };
-      ERR error = Action(AC(-5), this, &args);
-      return error;
-   }
-
-   // Customised field getting
-
-   inline ERR getProcedure(std::string_view &Value) noexcept {
-      Value = this->Procedure;
-      return ERR::Okay;
-   }
-
-   inline ERR getCacheFile(std::string_view &Value) noexcept {
-      Value = this->CacheFile;
-      return ERR::Okay;
-   }
-
-   inline ERR getPath(std::string_view &Value) noexcept {
-      Value = this->Path;
-      return ERR::Okay;
-   }
-
-   inline ERR getErrorMessage(std::string_view &Value) noexcept {
-      Value = this->ErrorMessage;
-      return ERR::Okay;
-   }
-
-   inline ERR getStatement(std::string_view &Value) noexcept {
-      Value = this->Statement;
-      return ERR::Okay;
-   }
-
-   inline ERR getWorkingPath(std::string_view &Value) noexcept {
-      auto field = &this->Class->Dictionary[8];
-      SetObjectContext(this, field, AC::NIL);
-      auto get_field = (ERR (*)(APTR, std::string_view &))field->GetValue;
-      auto error = get_field(this, Value);
-      RestoreObjectContext();
-      return error;
-   }
-
-   inline ERR getResults(std::span<std::string> &Value) noexcept {
-      auto ktv = (kt::vector<std::string> *)(((int8_t *)this) + 280);
-      Value = std::span<std::string>(ktv->data(), ktv->size());
-      return ERR::Okay;
-   }
-
-   inline ERR getTarget(OBJECTID &Value) noexcept {
-      Value = this->TargetID;
-      return ERR::Okay;
-   }
-
-   inline ERR getFlags(SCF &Value) noexcept {
-      Value = this->Flags;
-      return ERR::Okay;
-   }
-
-   inline ERR getError(ERR &Value) noexcept {
-      Value = this->Error;
-      return ERR::Okay;
-   }
-
-   inline ERR getCurrentLine(int &Value) noexcept {
-      Value = this->CurrentLine;
-      return ERR::Okay;
-   }
-
-   inline ERR getLineOffset(int &Value) noexcept {
-      Value = this->LineOffset;
-      return ERR::Okay;
-   }
-
-
-   // Customised field setting
-
-   inline ERR setProcedure(const std::string_view &Value) noexcept {
-      this->Procedure = Value;
-      return ERR::Okay;
-   }
-
-   inline ERR setCacheFile(const std::string_view &Value) noexcept {
-      this->CacheFile = Value;
-      return ERR::Okay;
-   }
-
-   inline ERR setPath(const std::string_view &Value) noexcept {
-      auto field = &this->Class->Dictionary[5];
-      return field->WriteValue(this, field, 0x00804500, &Value);
-   }
-
-   inline ERR setErrorMessage(const std::string_view &Value) noexcept {
-      this->ErrorMessage = Value;
-      return ERR::Okay;
-   }
-
-   inline ERR setStatement(const std::string_view &Value) noexcept {
-      auto field = &this->Class->Dictionary[11];
-      return field->WriteValue(this, field, 0x00804300, &Value);
-   }
-
-   inline ERR setWorkingPath(const std::string_view &Value) noexcept {
-      this->WorkingPath = Value;
-      return ERR::Okay;
-   }
-
-   inline ERR setResults(const kt::vector<std::string> &Value) noexcept {
-      this->Results = Value;
-      return ERR::Okay;
-   }
-
-   inline ERR setTarget(OBJECTID Value) noexcept {
-      this->TargetID = Value;
-      return ERR::Okay;
-   }
-
-   inline ERR setFlags(const SCF Value) noexcept {
-      if (this->initialised()) return ERR::ImmutableField;
-      this->Flags = Value;
-      return ERR::Okay;
-   }
-
-   inline ERR setLineOffset(const int Value) noexcept {
-      this->LineOffset = Value;
-      return ERR::Okay;
-   }
-
-};
-
-namespace sc {
-inline ERR Call(const FUNCTION &Function, std::span<const ScriptArg> Args) noexcept {
-   struct Callback args = { Function.ProcedureID, Args.data(), int(Args.size()), ERR::Okay };
-   return Action(sc::Callback::id, Function.Context, &args);
-}
-
-inline ERR Call(const FUNCTION &Function, std::span<const ScriptArg> Args, ERR &Result) noexcept {
-   struct Callback args = { Function.ProcedureID, Args.data(), int(Args.size()), ERR::Okay };
-   ERR error = Action(sc::Callback::id, Function.Context, &args);
-   Result = args.Error;
-   return(error);
-}
-
-template <std::size_t SIZE> ERR Call(const FUNCTION &Function, const std::array<ScriptArg, SIZE> &Args) noexcept {
-   return Call(Function, std::span<const ScriptArg>(Args));
-}
-
-template <std::size_t SIZE> ERR Call(const FUNCTION &Function, const std::array<ScriptArg, SIZE> &Args, ERR &Result) noexcept {
-   return Call(Function, std::span<const ScriptArg>(Args), Result);
-}
-
-inline ERR Call(const FUNCTION &Function) noexcept {
-   return Call(Function, std::span<const ScriptArg>());
-}
-
-inline ERR Call(const FUNCTION &Function, ERR &Result) noexcept {
-   return Call(Function, std::span<const ScriptArg>(), Result);
-}
-} // namespace
 struct ActionEntry {
    ERR (*PerformAction)(OBJECTPTR, APTR);     // Pointer to a custom action hook.
 };
