+++ include/kotuku/system/registry.h
@@ -10,7 +10,6 @@ enum class CLASSID : uint32_t {
    SOUND = 0x825a6b79UL,
    BITMAP = 0xff557b59UL,
    DISPLAY = 0x39413236UL,
-   REGION = 0x471e889aUL,
    PCX = 0x79dd2460UL,
    FILE = 0x4ca5c5b3UL,
    MP3 = 0xd961bed4UL,
@@ -19,7 +18,6 @@ enum class CLASSID : uint32_t {
    METACLASS = 0x883d1872UL,
    MODULE = 0xcd79e511UL,
    TIME = 0x0904cf78UL,
-   TIMER = 0x39967fa6UL,
    FONT = 0x58644726UL,
    MUSIC = 0x8729ca30UL,
    SCRIPT = 0xd44abc36UL,
@@ -35,9 +33,7 @@ enum class CLASSID : uint32_t {
    DOCUMENT = 0x9989342dUL,
    XML = 0x6293d85fUL,
    COMPRESSION = 0x2aebea46UL,
-   FTP = 0x17bfcff6UL,
    NETSERVER = 0x61231fb4UL,
-   WINDOW = 0x7fdabc89UL,
    BLUETOOTH = 0x79f3cdd0UL,
    NETSOCKET = 0xb58a423bUL,
    VIDEO = 0x7bd287f9UL,
@@ -52,13 +48,11 @@ enum class CLASSID : uint32_t {
    PROXY = 0x8dcea850UL,
    USBDEVICE = 0xc7654d55UL,
    CRYPT = 0xc9f07d81UL,
-   LAYOUT = 0x9ad45281UL,
    NETLOOKUP = 0xe6f0a23cUL,
    VECTOR = 0xa0b8f38aUL,
    JSON = 0x6b318546UL,
    SVG = 0xbf531772UL,
    RSVG = 0xbdd0261fUL,
-   SERIAL = 0xf2cd2b88UL,
    VECTORRECTANGLE = 0x62e3f082UL,
    VECTORELLIPSE = 0x0da60ceeUL,
    VECTORPATH = 0xecf78098UL,
