+++ src/audio/audio.cpp
@@ -95,6 +95,8 @@ For optimal results, choose the appropriate interface based on application requi
 
 #include <kotuku/main.h>
 #include <kotuku/modules/audio.h>
+#include <kotuku/modules/config.h>
+#include <kotuku/modules/script.h>
 #include <kotuku/strings.hpp>
 #include <sstream>
 #include <algorithm>
