+++ src/core/CMakeLists.txt
@@ -20,7 +20,6 @@ set (FUNCTIONS  # Full paths are required for add_custom_command() dependencies 
    "${CMAKE_CURRENT_SOURCE_DIR}/fs_volumes.cpp")
 
 set (CORE_CLASSES
-   "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_config.cpp"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_file.cpp"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_metaclass.cpp"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_module.cpp"
@@ -33,9 +32,7 @@ set (CORE_CLASSES
 
 SET (CORE_DEFS
    "${CMAKE_CURRENT_SOURCE_DIR}/compression/class_compressed_stream_def.c"
-   "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_config_def.c"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_file_def.c"
-   "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_script_def.c"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_task_def.c"
    "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_thread_def.c"
    "${CMAKE_CURRENT_SOURCE_DIR}/compression/class_compression_def.c")
@@ -47,7 +44,6 @@ if (KOTUKU_DEFS_AVAILABLE)
       "${DOCS}/modules/core.xml"
       "${DOCS}/modules/classes/compressedstream.xml"
       "${DOCS}/modules/classes/compression.xml"
-      "${DOCS}/modules/classes/config.xml"
       "${DOCS}/modules/classes/file.xml"
       "${DOCS}/modules/classes/metaclass.xml"
       "${DOCS}/modules/classes/module.xml"
@@ -89,6 +85,26 @@ if (KOTUKU_DEFS_AVAILABLE)
 
    register_generated_header_target (core_defs)
 endif ()
+
+idl_gen ("defs/config.tdl"
+   NAME config_defs
+   OUTPUT "${INCLUDE_OUTPUT}/modules/config.h"
+   BYPRODUCTS
+      "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_config_def.c"
+      "${DOCS}/modules/classes/config.xml"
+   DEPENDS
+      "${CMAKE_CURRENT_SOURCE_DIR}/defs/common.tdl"
+      "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_config.cpp")
+
+idl_gen ("defs/script.tdl"
+   NAME script_defs
+   OUTPUT "${INCLUDE_OUTPUT}/modules/script.h"
+   BYPRODUCTS
+      "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_script_def.c"
+      "${DOCS}/modules/classes/script.xml"
+   DEPENDS
+      "${CMAKE_CURRENT_SOURCE_DIR}/defs/common.tdl"
+      "${CMAKE_CURRENT_SOURCE_DIR}/classes/class_script.cpp")
 
 if (NOT MSVC)
    # A bug in (MinGW?) GCC requires the following flag is applied to ensure that __packed__ works correctly.
