+++ src/core/defs/registry.tdl
@@ -7,7 +7,6 @@ header({ path="system/registry", copyright="Paul Manias 1996-2026" }, function()
     "SOUND",
     "BITMAP",
     "DISPLAY",
-    "REGION",
     "PCX",
     "FILE",
     "MP3",
@@ -16,7 +15,6 @@ header({ path="system/registry", copyright="Paul Manias 1996-2026" }, function()
     "METACLASS",
     "MODULE",
     "TIME",
-    "TIMER",
     "FONT",
     "MUSIC",
     "SCRIPT",
@@ -32,9 +30,7 @@ header({ path="system/registry", copyright="Paul Manias 1996-2026" }, function()
     "DOCUMENT",
     "XML",
     "COMPRESSION",
-    "FTP",
     "NETSERVER",
-    "WINDOW",
     "BLUETOOTH",
     "NETSOCKET",
     "VIDEO",
@@ -49,13 +45,11 @@ header({ path="system/registry", copyright="Paul Manias 1996-2026" }, function()
     "PROXY",
     "USBDEVICE",
     "CRYPT",
-    "LAYOUT",
     "NETLOOKUP",
     "VECTOR",
     "JSON",
     "SVG",
     "RSVG",
-    "SERIAL",
     "VECTORRECTANGLE",
     "VECTORELLIPSE",
     "VECTORPATH",
