+++ src/core/defs/script.tdl
@@ -1,0 +1,103 @@
+--$TIRI:Include
+
+header({ module="Core", copyright="Paul Manias © 2026", timestamp=20240611 }, function()
+  c_include('<kotuku/main.h>')
+
+  restrict(function()
+    loadFile('./common.tdl')
+  end)
+
+  privateNames({ "ScriptArg" })
+
+  c_insert([[
+struct ScriptArg { // For use with sc::Exec
+   CSTRING Name;
+   uint32_t Type;
+   union {
+      APTR    Address;
+      int     Int;
+      int64_t Int64;
+      double  Double;
+   };
+
+   ScriptArg(CSTRING pName, OBJECTPTR pValue, uint32_t pType = FD_OBJECTPTR) : Name(pName), Type(pType), Address((APTR)pValue) { }
+   ScriptArg(CSTRING pName, std::string &pValue, uint32_t pType = FD_STRING) : Name(pName), Type(pType), Address((APTR)pValue.data()) { }
+   ScriptArg(CSTRING pName, const std::string &pValue, uint32_t pType = FD_STRING) : Name(pName), Type(pType), Address((APTR)pValue.data()) { }
+   ScriptArg(CSTRING pName, CSTRING pValue, uint32_t pType = FD_STRING) : Name(pName), Type(pType), Address((APTR)pValue) { }
+   ScriptArg(CSTRING pName, APTR pValue, uint32_t pType = FD_PTR) : Name(pName), Type(pType), Address(pValue) { }
+   ScriptArg(CSTRING pName, int pValue, uint32_t pType = FD_INT) : Name(pName), Type(pType), Int(pValue) { }
+   ScriptArg(CSTRING pName, uint32_t pValue, uint32_t pType = FD_INT) : Name(pName), Type(pType), Int(pValue) { }
+   ScriptArg(CSTRING pName, int64_t pValue, uint32_t pType = FD_INT64) : Name(pName), Type(pType), Int64(pValue) { }
+   ScriptArg(CSTRING pName, double pValue, uint32_t pType = FD_DOUBLE) : Name(pName), Type(pType), Double(pValue) { }
+};
+]])
+
+  methods("Script", "Sc", {
+    { id=1, name="Exec" },
+    { id=2, name="DerefProcedure" },
+    { id=3, name="Callback" },
+    { id=4, name="GetProcedureID" },
+    { id=5, name="DebugLog" }
+  })
+
+  klass("Script", { src={ "../classes/class_script.cpp" }, output="../classes/class_script_def.c" }, [[
+    string Procedure    # Specifies a procedure name to be executed
+    string CacheFile
+    string Path         # Location of the script
+    string ErrorMessage # Human readable error string defined on failure
+    string Statement
+    string WorkingPath
+    vector(string) Results
+    oid Target          # The object that script objects must be initialised to, e.g. for obj.new()
+    int(SCF) Flags      # Optional flags
+    error Error         # If an error occurred, this field will indicate the error number
+    int CurrentLine     # Current line being executed, or failed line if script execution terminated
+    int LineOffset      # An optional offset to use when reporting line numbers
+  ]],
+  nil,
+  [[
+   int64_t  ProcedureID;          // For callbacks
+   KEYVALUE Vars;                 // Global parameters
+   const ScriptArg *ProcArgs;     // Procedure args - applies during Exec
+   int      ActivationCount;      // Incremented every time the script is activated.
+   int      TotalArgs;            // Total number of ProcArgs
+   OBJECTID ScriptOwnerID;
+
+#ifdef KOTUKU_CXX_REUSES_BASE_TAIL_PADDING // Padding for the alignment of derived classes
+   uint8_t TailPadding[alignof(APTR) - sizeof(OBJECTID)];
+#endif
+  ]])
+
+  c_insert([==[
+namespace sc {
+inline ERR Call(const FUNCTION &Function, std::span<const ScriptArg> Args) noexcept {
+   struct Callback args = { Function.ProcedureID, Args.data(), int(Args.size()), ERR::Okay };
+   return Action(sc::Callback::id, Function.Context, &args);
+}
+
+inline ERR Call(const FUNCTION &Function, std::span<const ScriptArg> Args, ERR &Result) noexcept {
+   struct Callback args = { Function.ProcedureID, Args.data(), int(Args.size()), ERR::Okay };
+   ERR error = Action(sc::Callback::id, Function.Context, &args);
+   Result = args.Error;
+   return(error);
+}
+
+template <std::size_t SIZE> ERR Call(const FUNCTION &Function, const std::array<ScriptArg, SIZE> &Args) noexcept {
+   return Call(Function, std::span<const ScriptArg>(Args));
+}
+
+template <std::size_t SIZE> ERR Call(const FUNCTION &Function, const std::array<ScriptArg, SIZE> &Args, ERR &Result) noexcept {
+   return Call(Function, std::span<const ScriptArg>(Args), Result);
+}
+
+inline ERR Call(const FUNCTION &Function) noexcept {
+   return Call(Function, std::span<const ScriptArg>());
+}
+
+inline ERR Call(const FUNCTION &Function, ERR &Result) noexcept {
+   return Call(Function, std::span<const ScriptArg>(), Result);
+}
+} // namespace
+  ]==])
+
+end)
