+++ src/document/defs/document.tdl
@@ -5,7 +5,8 @@ module({ name="Document", copyright="Paul Manias © 2005-2026", version=1.0, tim
     loadFile(glPath .. 'common-graphics.tdl')
   end)
 
-  c_include("<kotuku/modules/display.h>", "<kotuku/modules/xml.h>", "<kotuku/modules/font.h>", "<kotuku/modules/vector.h>")
+  c_include("<kotuku/modules/display.h>", "<kotuku/modules/xml.h>", "<kotuku/modules/font.h>",
+    "<kotuku/modules/vector.h>", "<kotuku/modules/script.h>")
 
   loadFile(glPath .. 'common.tdl')
