+++ src/network/network.cpp
@@ -28,6 +28,7 @@ sockets and HTTP, please refer to the @NetSocket, @NetServer and @HTTP classes.
 
 #include <kotuku/main.h>
 #include <kotuku/modules/network.h>
+#include <kotuku/modules/script.h>
 #include <kotuku/strings.hpp>
 
 #include "net_platform.h"
