+++ src/origo/origo.cpp
@@ -12,6 +12,7 @@ This version of the launcher is intended for use from the command-line only.
 #include <kotuku/main.h>
 #include <kotuku/modules/core.h>
 #include <kotuku/modules/display.h>
+#include <kotuku/modules/script.h>
 #include <kotuku/startup.h>
 #include <kotuku/strings.hpp>
 #include <string.h>
