+++ src/vector/vector.h
@@ -27,6 +27,7 @@ template<class... Args> void DBG_TRANSFORM(Args...) {
 #include <kotuku/modules/image.h>
 #include <kotuku/modules/display.h>
 #include <kotuku/modules/font.h>
+#include <kotuku/modules/script.h>
 #include <kotuku/strings.hpp>
 
 using namespace kt;
