+++ src/vector/vector.tdl
@@ -3,7 +3,7 @@
 module({ name='Vector', copyright='Paul Manias © 2010-2026', version=1.0, timestamp=20240611,
     src={ 'vector_functions.cpp', 'vector_readpainter.cpp' }
   }, function()
-  c_include('<kotuku/modules/display.h>', '<kotuku/modules/image.h>')
+  c_include('<kotuku/modules/display.h>', '<kotuku/modules/image.h>', '<kotuku/modules/config.h>')
 
   restrict(function()
     loadFile(glPath .. 'common.tdl')
